Move `shuffle_extrinsics` from utils to domain_block_preprocessor
